Move open_cptdataarray to cptio
